+++ src/app/chat/features/ChatInput/ActionBar/config.ts
@@ -1,6 +1,7 @@
 import { FC } from 'react';
 
 import Clear from './Clear';
+import FileUpload from './FileUpload';
 import History from './History';
 import ModelSwitch from './ModelSwitch';
 import Temperature from './Temperature';
@@ -8,6 +9,7 @@ import Token from './Token';
 
 export const actionMap: Record<string, FC> = {
   clear: Clear,
+  fileUpload: FileUpload,
   history: History,
   model: ModelSwitch,
   temperature: Temperature,
@@ -15,5 +17,6 @@ export const actionMap: Record<string, FC> = {
 };
 
 // we can make these action lists configurable in the future
-export const leftActionList = ['model', 'temperature', 'history'];
+export const leftActionList = ['model', 'temperature', 'fileUpload', 'history'];
+
 export const rightActionList = ['clear'];