+++ src/services/file.ts
@@ -1,0 +1,36 @@
+import { FileModel } from '@/database/models/file';
+import { LocalFile } from '@/types/database/files';
+import { FilePreview } from '@/types/files';
+
+class FileService {
+  async uploadFile(file: LocalFile) {
+    // save to local storage
+    // we may want to save to a remote server later
+    return FileModel.create(file);
+  }
+
+  async removeFile(id: string) {
+    return FileModel.delete(id);
+  }
+
+  async getFile(id: string): Promise<FilePreview> {
+    const item = await FileModel.findById(id);
+    if (!item) {
+      throw new Error('file not found');
+    }
+
+    // arrayBuffer to url
+    const url = URL.createObjectURL(new Blob([item.data]));
+    const base64 = Buffer.from(item.data).toString('base64');
+
+    return {
+      base64Url: `data:${item.fileType};base64,${base64}`,
+      fileType: item.fileType,
+      name: item.name,
+      saveMode: 'local',
+      url,
+    };
+  }
+}
+
+export const fileService = new FileService();