+++ src/store/files/initialState.ts
@@ -1,0 +1,7 @@
+import { ImageFileState, initialImageFileState } from './slices/images';
+
+export type FilesStoreState = ImageFileState;
+
+export const initialState: FilesStoreState = {
+  ...initialImageFileState,
+};