+++ src/store/files/slices/images/initialState.ts
@@ -1,0 +1,11 @@
+import { FilePreview } from '@/types/files';
+
+export interface ImageFileState {
+  imagesMap: Record<string, FilePreview>;
+  inputFilesList: string[];
+}
+
+export const initialImageFileState: ImageFileState = {
+  imagesMap: {},
+  inputFilesList: [],
+};