+++ src/types/database/db.ts
@@ -1,0 +1,4 @@
+export type DBModel<T> = T & {
+  createdAt: number;
+  id: string;
+};