+++ src/types/openai/chat.ts
@@ -1,12 +1,26 @@
 import { OpenAIFunctionCall } from '@/types/chatMessage';
 import { LLMRoleType } from '@/types/llm';
 
+interface UserMessageContentPartText {
+  text: string;
+  type: 'text';
+}
+interface UserMessageContentPartImage {
+  image_url: {
+    detail?: 'auto' | 'low' | 'high';
+    url: string;
+  };
+  type: 'image_url';
+}
+
+export type UserMessageContentPart = UserMessageContentPartText | UserMessageContentPartImage;
+
 export interface OpenAIChatMessage {
   /**
    * @title 内容
    * @description 消息内容
    */
-  content: string;
+  content: string | UserMessageContentPart[];
 
   function_call?: OpenAIFunctionCall;
   name?: string;